moved recipes to conform to new spack repo layout
As explained here:
https://spack.readthedocs.io/en/v1.1.1/repositories.html#package-repositories-repos-yaml
