+++ spack_repo/lihpc/meshing/repo.yaml
@@ -1,0 +1,3 @@
+repo:
+  namespace: lihpc.meshing
+  api: v2.2
