+++ spack_repo/lihpc/meshing_supersede/packages/vtk_maillage/vtk-maillage_gcc11_const.patch
