+++ spack_repo/lihpc/meshing_supersede/repo.yaml
@@ -1,0 +1,3 @@
+repo:
+  namespace: lihpc.meshing_supersede
+  api: v2.2
